[ refactor ] 팀 단위 코드 리팩토링
