+++ package.json
@@ -1,33 +1,35 @@
 {
-  "name": "next-js",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "next": "15.4.6",
-    "react": "19.1.0",
-    "react-dom": "19.1.0"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "@typescript-eslint/eslint-plugin": "^8.31.1",
-    "@typescript-eslint/parser": "^8.31.1",
-    "eslint": "^9",
-    "eslint-config-next": "15.4.6",
-    "eslint-config-prettier": "^9.1.2",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier-eslint": "^16.1.2",
-    "prettier": "^3.2.5",
-    "tailwindcss": "^4",
-    "typescript": "^5"
-  }
+    "name": "next-js",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev --turbopack",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@tanstack/react-query": "^5.85.5",
+        "next": "15.4.6",
+        "react": "19.1.0",
+        "react-dom": "19.1.0"
+    },
+    "devDependencies": {
+        "@eslint/eslintrc": "^3",
+        "@tailwindcss/postcss": "^4",
+        "@types/node": "^20",
+        "@types/react": "^19",
+        "@types/react-dom": "^19",
+        "@typescript-eslint/eslint-plugin": "^8.31.1",
+        "@typescript-eslint/parser": "^8.31.1",
+        "axios": "^1.11.0",
+        "eslint": "^9",
+        "eslint-config-next": "15.4.6",
+        "eslint-config-prettier": "^9.1.2",
+        "eslint-plugin-prettier": "^5.1.3",
+        "prettier": "^3.2.5",
+        "prettier-eslint": "^16.1.2",
+        "tailwindcss": "^4",
+        "typescript": "^5"
+    }
 }
