+++ src/app/search/_types/book.ts
@@ -1,0 +1,9 @@
+interface Book {
+    id: number;
+    title: string;
+    author: string;
+    coverImage: string;
+    price: number;
+}
+
+export type { Book };
