+++ src/providers/QueryProvider.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { useState } from 'react';
+
+export function QueryProvider({ children }: { children: React.ReactNode }) {
+    const [queryClient] = useState(
+        () =>
+            new QueryClient({
+                defaultOptions: {
+                    queries: {
+                        staleTime: 5 * 60 * 1000, // 5분
+                        gcTime: 10 * 60 * 1000, // 10분 (구 cacheTime)
+                    },
+                },
+            }),
+    );
+
+    return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
+}
